format docs
