+++ docs/gradsflow/callbacks.md
@@ -1,1 +1,5 @@
-::: gradsflow.callbacks
+Callbacks
+
+::: gradsflow.callbacks.comet.CometCallback
+::: gradsflow.callbacks.wandb.WandbCallback
+::: gradsflow.callbacks.gpu.EmissionTrackerCallback
