+++ gradsflow/callbacks/logger.py
@@ -11,21 +11,9 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
+
 import os
 from pathlib import Path
-from typing import Optional
 
 import pandas as pd
 from loguru import logger
@@ -71,29 +59,3 @@ class CSVLogger(Callback):
         self._logs.append(data)
         df = pd.DataFrame(self._logs)
         df.to_csv(self._dst, index=False)
-
-
-class ModelCheckpoint(Callback):
-    """
-    Saves Model checkpoint
-    Args:
-        filename: name of checkpoint
-        path: folder path location of the model checkpoint. Will create a folder if does not exist.
-        save_extra: whether to save extra details like tracker
-    """
-
-    _name = "ModelCheckpoint"
-
-    def __init__(self, filename: Optional[str] = None, path: str = os.getcwd(), save_extra: bool = False):
-        super().__init__(model=None)
-        filename = Path(filename or "model")
-        path = Path(path)
-        self.path = path
-        self.path.mkdir(exist_ok=True)
-        self._dst = path / filename
-        self.save_extra = save_extra
-
-    def on_epoch_end(self):
-        epoch = self.model.tracker.current_epoch
-        path = f"{self._dst}_epoch={epoch}_.pt"
-        self.model.save(path, save_extra=self.save_extra)
