+++ gradsflow/callbacks/wandb.py
@@ -23,7 +23,7 @@ from gradsflow.utility.imports import requires
 CURRENT_FILE = os.path.dirname(os.path.realpath(__file__))
 
 
-def define_metrics():
+def define_metrics(default_step_metric: str = "global_step"):
     min_max_def: Dict[str, List[str]] = {
         "min": ["train/step_loss", "train/epoch_loss", "val/epoch_loss"],
         "max": ["train/acc*", "val/acc*"],
@@ -32,23 +32,35 @@ def define_metrics():
         for metric in metric_list:
             if "epoch" in metric or "val" in metric:
                 wandb.define_metric(metric, summary=summary, step_metric="epoch")
-    wandb.define_metric("*", step_metric="global_step")
+    wandb.define_metric("*", step_metric=default_step_metric)
 
 
 class WandbCallback(Callback):
     """
     [Weights & Biases](https://www.wandb.com/) Logging callback. To use this callback `pip install wandb`.
+    Any metric that contains `epoch` will be plotted with `epoch` and all the other metrics will be plotted against
+    `global_step` which is total training steps. You can change the default axis by providing `default_step_metric`.
+
     Args:
         log_model: Whether to upload model artifact to Wandb
         code_file: path of the code you want to upload as artifact to Wandb
+        default_step_metric: Metrics will be plotted against the `default_step_metric`. Default value is `global_step`.
+
+    ```python
+    from gradsflow.callbacks import WandbCallback
+    from timm import create_model
+
+    cnn = create_model("resnet18", pretrained=False, num_classes=1)
+    model = Model(cnn)
+    model.compile()
+    cb = WandbCallback()
+    autodataset = None  # create your dataset
+    model.fit(autodataset, callbacks=cb)
+    ```
     """
 
     @requires("wandb", "WandbCallback requires wandb to be installed!")
-    def __init__(
-        self,
-        log_model: bool = False,
-        code_file: Optional[str] = None,
-    ):
+    def __init__(self, log_model: bool = False, code_file: Optional[str] = None, default_step_metric="global_step"):
         super().__init__()
         if wandb.run is None:
             raise ValueError("You must call wandb.init() before WandbCallback()")
@@ -56,10 +68,10 @@ class WandbCallback(Callback):
         self._train_prefix = "train"
         self._val_prefix = "val"
         self._log_model = log_model
-        self._setup()
+        self._setup(default_step_metric)
 
-    def _setup(self):
-        define_metrics()
+    def _setup(self, default_step_metric):
+        define_metrics(default_step_metric)
 
     def on_fit_start(self):
         if self._log_model:
